회원탈퇴 다시 구현함

+++ app/mod_main/controllers.py
@@ -1,15 +1,15 @@
-#-*- coding: utf-8 -*-
+# -*- coding: utf-8 -*-
 
 # Import flask dependencies
 from flask import Blueprint, render_template, redirect, url_for, flash
 from flask.ext.login import login_required, current_user, logout_user
-from app.models import social, user_datastore, security
+from app.models import social, user_datastore
 from forms import UserLeaveForm
-from flask_security import roles_required
+from app import db
 from flask_security.utils import verify_and_update_password
 
 
-# Define the blueprint: 'main', set its url prefix: app.u   rl/main
+# Define the blueprint: 'main', set its url prefix: app.url/main
 mod_main = Blueprint('main', __name__, url_prefix='')
 
 
@@ -47,28 +47,30 @@ def profile():
 @mod_main.route('/leave', methods=['POST', 'GET'])
 @login_required
 def leave():
-    ispwcorrect = False
+    is_pw_correct = False
     form = UserLeaveForm()
     if form.validate_on_submit():
         if verify_and_update_password(form.password.data, current_user):
             flash("Your input is correct.")
-            ispwcorrect = True
-            user_datastore.create_role(name='leave', description='This user will be leave')
-            user_datastore.add_role_to_user(current_user, 'leave')
-            user_datastore.commit()
-
-            render_template('main/leave.html', form=form, ispwcorrect=ispwcorrect)
+            is_pw_correct = True
+            current_user.is_want_leave = True
+            db.session.commit()
+            return render_template('main/leave.html', form=form, is_pw_correct=is_pw_correct)
         else:
             flash("Your input is not correct.")
             return redirect(url_for('main.leave'))
-    return render_template('main/leave.html', form=form, ispwcorrect=ispwcorrect)
+    return render_template('main/leave.html', form=form, is_pw_correct=is_pw_correct)
 
 
 @mod_main.route('/delete', methods=['POST', 'GET'])
 @login_required
-@roles_required('leave')
 def deleteuser():
-    user_datastore.delete_user(current_user)
-    logout_user()
-    user_datastore.commit()
-    return redirect(url_for('main.index'))
+    if current_user.is_want_leave:
+        user_datastore.delete_user(current_user)
+        logout_user()
+        user_datastore.commit()
+        flash('Your account is successfully deleted!')
+        return redirect(url_for('main.index'))
+    else:
+        flash('It is wrong approach!!!')
+        return render_template('404.html')
